pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cue-agent"
version = "0.1.0"
description = "Autonomous AI agent built on the Efficient Agent Protocol"
requires-python = ">=3.11"
dependencies = [
    "efficient-agent-protocol @ git+https://github.com/GenieWeenie/efficient-agent-protocol.git",
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "python-telegram-bot[webhooks]>=22.0",
    "apscheduler>=4.0.0a5",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]
vector = [
    "chromadb>=0.5.0",
]

[project.scripts]
cue-agent = "cue_agent.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
files = ["src/cue_agent"]
strict = true
ignore_missing_imports = true
warn_return_any = true
